+++ src/main/java/janggi/domain/piece/General.java
@@ -1,7 +1,7 @@
 package janggi.domain.piece;
 
 import janggi.domain.Dynasty;
-import janggi.domain.piece.movement.MovePath;
+import janggi.domain.piece.movepath.MovePath;
 import java.util.Set;
 
 public class General extends Piece {
