+++ src/main/java/janggi/domain/piece/Piece.java
@@ -2,7 +2,7 @@ package janggi.domain.piece;
 
 import janggi.domain.Dynasty;
 import janggi.domain.board.Point;
-import janggi.domain.piece.movement.MovePath;
+import janggi.domain.piece.movepath.MovePath;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
