+++ src/test/java/janggi/domain/piece/CannonTest.java
@@ -12,17 +12,16 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class HanSoldierTest {
-
-    @DisplayName("한나라 졸이 목적지까지 가는데 거치는 포인트를 알 수 있다")
+class CannonTest {
+    @DisplayName("포가 목적지까지 가는데 거치는 포인트를 알 수 있다")
     @ParameterizedTest
     @MethodSource("provideMovablePosition")
     void movePath(Point from, Point to, List<Point> expected) {
         //given
-        HanSoldier soldier = new HanSoldier();
+        Chariot ch = new Chariot(Dynasty.HAN);
 
         //when
-        List<Point> points = soldier.movePath(from, to);
+        List<Point> points = ch.movePath(from, to);
 
         //then
         assertThat(points).isEqualTo(expected);
@@ -33,10 +32,10 @@ class HanSoldierTest {
     @MethodSource("providePiecesOnPath")
     void canMove(PiecesOnPath piecesOnPath, boolean expected) {
         //given
-        HanSoldier soldier = new HanSoldier();
+        Chariot horse = new Chariot(Dynasty.HAN);
 
         //when
-        boolean actual = soldier.canMove(piecesOnPath);
+        boolean actual = horse.canMove(piecesOnPath);
 
         //then
         assertThat(actual).isEqualTo(expected);
@@ -44,9 +43,12 @@ class HanSoldierTest {
 
     private static Stream<Arguments> provideMovablePosition() {
         return Stream.of(
-                Arguments.of(Fixtures.SIX_FIVE, Fixtures.SIX_FOUR, List.of(Fixtures.SIX_FOUR)),
-                Arguments.of(Fixtures.SIX_FIVE, Fixtures.SIX_SIX, List.of(Fixtures.SIX_SIX)),
-                Arguments.of(Fixtures.SIX_FIVE, Fixtures.SEVEN_FIVE, List.of(Fixtures.SEVEN_FIVE))
+                Arguments.of(Fixtures.SIX_FIVE, Fixtures.FOUR_FIVE, List.of(Fixtures.FIVE_FIVE, Fixtures.FOUR_FIVE)),
+                Arguments.of(Fixtures.SIX_FIVE, Fixtures.SIX_ONE,
+                        List.of(Fixtures.SIX_FOUR, Fixtures.SIX_THREE, Fixtures.SIX_TWO, Fixtures.SIX_ONE)),
+                Arguments.of(Fixtures.SIX_FIVE, Fixtures.SIX_NINE,
+                        List.of(Fixtures.SIX_SIX, Fixtures.SIX_SEVEN, Fixtures.SIX_EIGHT, Fixtures.SIX_NINE)),
+                Arguments.of(Fixtures.SIX_FIVE, Fixtures.EIGHT_FIVE, List.of(Fixtures.SEVEN_FIVE, Fixtures.EIGHT_FIVE))
         );
     }
