+++ src/test/java/janggi/domain/piece/ChuSoldierTest.java
@@ -14,15 +14,15 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 class ChuSoldierTest {
 
-    @DisplayName("한나라 졸이 목적지까지 가는데 거치는 포인트를 알 수 있다")
+    @DisplayName("추나라 졸이 목적지까지 가는데 거치는 포인트를 알 수 있다")
     @ParameterizedTest
     @MethodSource("provideMovablePosition")
     void movePath(Point from, Point to, List<Point> expected) {
         //given
-        ChuSoldier horse = new ChuSoldier();
+        ChuSoldier soldier = new ChuSoldier();
 
         //when
-        List<Point> points = horse.movePath(from, to);
+        List<Point> points = soldier.movePath(from, to);
 
         //then
         assertThat(points).isEqualTo(expected);
@@ -33,10 +33,10 @@ class ChuSoldierTest {
     @MethodSource("providePiecesOnPath")
     void canMove(PiecesOnPath piecesOnPath, boolean expected) {
         //given
-        ChuSoldier horse = new ChuSoldier();
+        ChuSoldier soldier = new ChuSoldier();
 
         //when
-        boolean actual = horse.canMove(piecesOnPath);
+        boolean actual = soldier.canMove(piecesOnPath);
 
         //then
         assertThat(actual).isEqualTo(expected);
